Modified accounts model so data object to accept Decimals as balance data type

+++ src/models/account.rs
@@ -1,17 +1,15 @@
-// Accounts models.rs
 use chrono::{DateTime, Utc};
 use serde::{Deserialize, Serialize};
-use sqlx::FromRow;
+use sqlx::{FromRow, types::BigDecimal};
 use uuid::Uuid;
-
 
 #[derive(Debug, FromRow, Serialize, Deserialize, Clone)]
 pub struct Account {
     pub id: Uuid,
     pub name: String,
     #[sqlx(rename = "type")]
-    pub type_: String,  // e.g., 'asset', 'liability', etc.
-    pub balance: f64,
+    pub type_: String,
+    pub balance: BigDecimal,  // ← Changed from f64 (or rust_decimal::Decimal)
     pub user_id: Uuid,
     pub created_at: DateTime<Utc>,
 }
